requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.0
python-dotenv>=1.0
